CRI: Move reference sorting to reference package
